# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MySchool API",
    "version": "2.2.0",
    "description": "API REST complète pour MySchool e-learning. \n        Gestion des utilisateurs, cours, chapitres, leçons, exercices, inscriptions et instructeurs.\n        Support complet des requêtes filtrées et index Firestore optimisés.",
    "contact": {
      "name": "MySchool Support",
      "email": "support@myschool.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-myschool-f862b.cloudfunctions.net",
      "description": "Serveur de production (Firebase Functions)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    }
  ],
...
  "paths": {
...
    "/api/subscriptions/create-payment": {
      "post": {
        "summary": "Créer un paiement Wave pour un abonnement à une catégorie",
        "tags": [
          "Subscriptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "plan",
                  "userId",
                  "category"
                ],
                "properties": {
                  "plan": {
                    "type": "string",
                    "enum": [
                      "MONTHLY",
                      "QUARTERLY",
                      "ANNUAL"
                    ],
                    "description": "Plan d'abonnement",
                    "example": "MONTHLY"
                  },
                  "userId": {
                    "type": "string",
                    "description": "ID de l'utilisateur",
                    "example": "VLzZCBlJstQOWhJy9Xg4MEO7ESK2"
                  },
                  "category": {
                    "type": "string",
                    "description": "Catégorie de cours à débloquer",
                    "example": "Mathématiques"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "paymentUrl": {
                          "type": "string",
                          "description": "URL de redirection vers Wave",
                          "example": "https://checkout.wave.com/pay/abc123"
                        },
                        "paymentId": {
                          "type": "string",
                          "example": "wave_12345"
                        },
                        "subscriptionId": {
                          "type": "string",
                          "example": "sub_abc123"
                        },
                        "amount": {
                          "type": "number",
                          "example": 3000
                        },
                        "plan": {
                          "type": "string",
                          "example": "MONTHLY"
                        },
                        "category": {
                          "type": "string",
                          "example": "Mathématiques"
                        },
                        "currency": {
                          "type": "string",
                          "example": "XOF"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Paiement d'abonnement créé pour la catégorie \"Mathématiques\""
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou abonnement actif existant pour cette catégorie"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          }
        }
      }
    },
    "/api/subscriptions/active/{userId}": {
      "get": {
        "summary": "Obtenir l'abonnement actif d'un utilisateur",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Abonnement actif trouvé"
          },
          "404": {
            "description": "Aucun abonnement actif"
          }
        }
      }
    },
    "/api/subscriptions/user/{userId}": {
      "get": {
        "summary": "Obtenir tous les abonnements d'un utilisateur",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des abonnements"
          }
        }
      }
    },
    "/api/subscriptions/{subscriptionId}/cancel": {
      "post": {
        "summary": "Annuler un abonnement",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "subscriptionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'abonnement"
          }
        ],
        "responses": {
          "200": {
            "description": "Abonnement annulé avec succès"
          }
        }
      }
    },
    "/api/subscriptions/check-access/{userId}": {
      "get": {
        "summary": "Vérifier les abonnements actifs d'un utilisateur",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de l'utilisateur"
          }
        ],
        "responses": {
          "200": {
            "description": "Statut d'accès avec liste des abonnements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "hasUnlimitedAccess": {
                          "type": "boolean",
                          "example": true
                        },
                        "activeSubscriptions": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "category": {
                                "type": "string",
                                "example": "Mathématiques"
                              },
                              "plan": {
                                "type": "string",
                                "example": "MONTHLY"
                              },
                              "endDate": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}